requirements.txt
easydict==1.9
Jinja2==3.0.3
python-dateutil==2.8.1
PyYAML==5.4.1
schema==0.7.4
stix2==2.1.0
